+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==================
+
+  * Combine overlapping ranges
+
 1.1.0 / 2016-05-13
 ==================